Returning the original failure when ErrorCollector has only collected a single failure
